tools/extraction-validator/requirements.txt
docling>=2.0
pyyaml>=6.0
